rename files | refact: .gitignore
