+++ src/enum/components/base/caseAlert.interface.ts
